Add the daemon flag to BaseConnection types with background thread
Several implementations of the BaseConnection class instantiate a
background thread to handle incoming data. At the moment, these threads
run without the daemon flag. With this configuration, the process will
not exit until all threads have exited. This means that the application
will hang indefinitely if an exception or natural exit condition causes
the main thread to end execution without shutting down the client first.
With this patch, the behavior is changed such that the threads created
in the BaseConnection instances will automatically stop once the main
thread exits.

This change affects the following classes:
* SocketConnection
* IsoTPSocketConnection
* PythonIsoTpConnection