Update ReadDataByIdentifier.py
When use rdbi and the DID constains multiple signals with the same data type (e.g. sint16), the offset and the val-tuple-concatenation have to be modified.
In my main file, the DID is configured with:
  config['data_identifiers'] = {'default': '>H', 0x4FD3: udsoncan.DidCodec('>h')}
The DID 0x4FD3 contains 2 signals of the same data type e.g. sint16.
Next step: find other method to handle the rdbi when DID contains multiple signals of the same data type, and how to handle it when the signals have different data types.